:recycle: [GDPRSubscriber] Add superclass in order to refacto
